+++ spark/v4.0/spark/src/main/java/org/apache/iceberg/spark/SortOrderToSpark.java
@@ -50,7 +50,7 @@ class SortOrderToSpark implements SortOrderVisitor<SortOrder> {
 
   @Override
   public SortOrder truncate(
-      String sourceName, int id, int width, SortDirection direction, NullOrder nullOrder) {
+      String sourceName, int id, long width, SortDirection direction, NullOrder nullOrder) {
     return Expressions.sort(
         Expressions.apply(
             "truncate", Expressions.literal(width), Expressions.column(quotedName(id))),
